requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
Pillow>=10.0.1
PyPDF2==3.0.1
google-cloud-documentai==2.9.0
google-auth==2.22.0
python-multipart==0.0.6
# Allow pip to choose a Pillow binary wheel compatible with your Python version
